refactor: 공유 버튼, 동기화 버튼 컴포넌트 및 스타일 분리
