+++ client/src/GalleryPage/components/ShareButton/ShareModal.scss
@@ -1,0 +1,27 @@
+.share-modal {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  height: 100%;
+  & > .button__container {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    & > button {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-size: 19px;
+      width: 100%;
+      height: 40px;
+      margin: 10px;
+      color: #ffffff;
+      background: #333333;
+      box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+      border-radius: 10px;
+      &.hidden {
+        display: none;
+      }
+    }
+  }
+}
