+++ client/src/GalleryPage/components/SyncButton/style.scss
@@ -1,0 +1,27 @@
+.create-modal {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  .make-gallery {
+    font-size: 30px;
+    margin-bottom: 20px;
+  }
+  & > button {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 19px;
+    width: 60%;
+    height: 50px;
+    color: #ffffff;
+    background: #333333;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 20px;
+    .fa {
+      margin-left: -12px;
+      margin-right: 8px;
+    }
+  }
+}
