+++ client/src/GalleryPage/components/ViewsCount/style.scss
@@ -1,0 +1,34 @@
+.share-view-count {
+  .share-span {
+    position: relative;
+    top: 10px;
+    left: -10px;
+    color: #ffffff;
+  }
+  .share-view-count {
+    text-align: left;
+    position: relative;
+    display: inline-block;
+    top: -10px;
+    left: 90px;
+    color: #ffffff;
+    width: 100px;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    transform: translate(-50%, 0);
+  }
+
+  [data-views]:hover::after {
+    content: attr(data-views);
+    position: absolute;
+    display: inline;
+    z-index: 999;
+    top: -20px;
+    background-color: #000000;
+    background-color: rgba(0, 0, 0, 0.8);
+    box-shadow: 0px 0px 3px 1px rgba(50, 50, 50, 0.4);
+    border-radius: 8px;
+    padding: 5px 10px;
+  }
+}
