+++ client/src/components/modal/style.scss
@@ -1,19 +1,28 @@
-.fullscreen-modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  padding: 20px;
-  border-radius: 10px;
-  background: #fff;
-  transform: translate(-50%, -50%);
-  z-index: 2;
+.fullscreen-modal-wrapper {
+  position: absolute;
+  &.hidden {
+    visibility: hidden;
+    pointer-events: none;
+  }
+  .fullscreen-modal {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    padding: 20px;
+    border-radius: 10px;
+    background: #fff;
+    transform: translate(-50%, -50%);
+    z-index: 2;
+  }
+  .dimmed {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(34, 34, 34, 0.3);
+    z-index: 1;
+  }
 }
-.dimmed {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(34, 34, 34, 0.3);
-  z-index: 1;
-}
+
+
